Reordering of the coupling calls in ROMS to insure that coupling information from CICE is used in the first time step.
Before this fix coupling was not done at the first time-step (iic
.eq. ntstart) in main3d.F. The calculation of the ocean state for the
next time step was missing information about the ice. Similar, at the
last time-step (iic .eq. ntend), the ROMS received coupling fields that
was not used (the model terminates in the middle of the time step),
before the new state is calculated.

In the fixed version the coupling in ROMS is done 1 time step earlier
than before. This seems to be different from what has been done with
atmosphere and wave coupling, and should be investigated further if
coupling to these compartments are an option.